refactor: excluindo arquivos individuais modMatricula para unificar em um único arquivo.
