--- src/modMatricula.h
@@ -1,6 +1,0 @@
-#ifndef MODULO_MATRICULA_H
-#define MODULO_MATRICULA_H
-
-char moduloMatricula(void);
-
-#endif
