--- src/ui/matricula/atualizarMatricula.h
@@ -1,6 +1,0 @@
-#ifndef ATUALIZAR_MATRICULA_H
-#define ATUALIZAR_MATRICULA_H
-
-char telaAtualizarMatricula(void);
-
-#endif
