--- src/ui/matricula/cadastrarMatricula.h
@@ -1,6 +1,0 @@
-#ifndef CADASTRAR_MATRICULA_H
-#define CADASTRAR_MATRICULA_H
-
-char telaCadastrarMatricula(void);
-
-#endif
