--- src/ui/matricula/excluirMatricula.c
@@ -1,18 +1,0 @@
-#include <stdio.h>
-
-char telaExcluirMatricula(void) {
-
-    char op;    
-
-    printf("====================\n");
-    printf("EXCLUIR MATRICULA\n");
-    printf("====================\n");
-    printf("  ...EM BREVE... \n");
-    printf("\n");
-    printf("<Aperte 0 para sair!>\n");
-
-    scanf("%c", &op);
-    getchar();
-
-    return op;
-}
