--- src/ui/matricula/excluirMatricula.h
@@ -1,6 +1,0 @@
-#ifndef EXCLUIR_MATRICULA_H
-#define EXCLUIR_MATRICULA_H
-
-char telaExcluirMatricula(void);
-
-#endif
