--- src/ui/matricula/telaMatricula.c
@@ -1,20 +1,0 @@
-#include <stdio.h>
-
-char telaMatricula(void) {
-
-    char op;
-
-    printf("====================\n");
-    printf("MATRICULAS\n");
-    printf("====================\n");
-    printf("(1) CADASTRAR\n");
-    printf("(2) VISUALIZAR\n");
-    printf("(3) ATUALIZAR\n");
-    printf("(4) EXCLUIR\n");
-    printf("(0) SAIR\n");
-
-    scanf("%c", &op);
-    getchar();
-
-    return op;
-}
