--- src/ui/matricula/telaMatricula.h
@@ -1,6 +1,0 @@
-#ifndef TELA_MATRICULA_H
-#define TELA_MATRICULA_H
-
-char telaMatricula(void);
-
-#endif
