--- src/ui/matricula/visualizarMatricula.c
@@ -1,18 +1,0 @@
-#include <stdio.h>
-
-char telaVisualizarMatricula(void) {
-
-    char op;
-    
-    printf("====================\n");
-    printf("VISUALIZAR MATRICULAS\n");
-    printf("====================\n");
-    printf("  ...EM BREVE... \n");
-    printf("\n");
-    printf("<Aperte 0 para sair!>\n");
-    
-    scanf("%c", &op);
-    getchar();
-
-    return op;
-    }
