--- src/ui/matricula/visualizarMatricula.h
@@ -1,6 +1,0 @@
-#ifndef VISUALIZAR_MATRICULA_H
-#define VISUALIZAR_MATRICULA_H
-
-char telaVisualizarMatricula(void);
-
-#endif
