+++ src/ui/telasMatricula.h
@@ -1,0 +1,11 @@
+#ifndef TELASMATRICULA_H
+#define TELASMATRICULA_H
+
+char moduloMatricula(void);
+char telaMatricula(void);
+char telaCadastrarMatricula(void);
+char telaVisualizarMatricula(void);
+char telaAtualizarMatricula(void);
+char telaExcluirMatricula(void);
+
+#endif
